+++ aerokit/aero/model1D.py
@@ -2,7 +2,7 @@
   representation of 1D flows
 """
 
-import math
+#import math
 import numpy     as np
 #import ShockWave as sw
 #import IterativeSolve
@@ -46,21 +46,28 @@ class state():
 		self.__init__(rho=gam * ps / a2, u=u, p=ps)
 
 	def compute_from_pt_rtt_p(self, pt, rtt, p):
+		"""Init state from Ptot r.Ttot and Ps (velocity sign is arbitrary and positive)
+
+		Args:
+			pt ([float]): [description]
+			rtt ([float]): [description]
+			p ([float]): [description]
+		"""
 		M   = Is.Mach_PtPs(pt/p, self._gamma)
 		rts = rtt/Is.TtTs_Mach(M, self._gamma)
 		self.__init__(rho=p/rts, u=M*np.sqrt(self._gamma*rts), p=p)
 		
 	def asound(self):
 		"""returns speed of sound"""
-		return math.sqrt(self._gamma*self.p/self.rho)
+		return np.sqrt(self._gamma*self.p/self.rho)
 
 	def left_acoustic(self):
 		"""returns left 'actual' speed of sound"""
-		return self.u-math.sqrt(self._gamma*self.p/self.rho)
+		return self.u-np.sqrt(self._gamma*self.p/self.rho)
 
 	def right_acoustic(self):
 		"""returns left 'actual' speed of sound"""
-		return self.u+math.sqrt(self._gamma*self.p/self.rho)
+		return self.u+np.sqrt(self._gamma*self.p/self.rho)
 
 	def Mach(self):
 		"""returns Mach number"""